number details goes to SQL but not PHP mailer
hit and miss, cant figure out why.  Fixed the numberdetails so it
submits Yes instead of 2 for all entries.  This was showing up on the
SMTP message as well.  Reconfigured add/remove as well as hide/show.  It
is more streamlined.  Test can be deleted now that it works. or it can
stay there and work as a test for all our others