Order Details Display
Sometimes it works, sometimes it doesn't  it can recognize the variables
when they are echoed, can only sometimes see them in html display.  So
far it seems when it does display in html it will get stuck on a single
order, for example the details for order 37, even if i click 2, 16, 22
etc.